+++ validator/client/src/test/java/tech/pegasys/teku/validator/client/AbstractDutySchedulerTest.java
@@ -13,26 +13,27 @@
 
 package tech.pegasys.teku.validator.client;
 
-import org.junit.jupiter.api.BeforeEach;
-import tech.pegasys.teku.bls.BLSPublicKey;
-import tech.pegasys.teku.core.signatures.Signer;
-import tech.pegasys.teku.datastructures.state.ForkInfo;
-import tech.pegasys.teku.datastructures.util.DataStructureUtil;
-import tech.pegasys.teku.infrastructure.async.StubAsyncRunner;
-import tech.pegasys.teku.metrics.StubMetricsSystem;
-import tech.pegasys.teku.validator.api.ValidatorApiChannel;
-import tech.pegasys.teku.validator.client.duties.AttestationProductionDuty;
-import tech.pegasys.teku.validator.client.duties.ValidatorDutyFactory;
-
-import java.util.Collection;
-import java.util.Optional;
-import java.util.Set;
-
 import static java.util.Collections.emptyList;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static tech.pegasys.teku.infrastructure.async.SafeFuture.completedFuture;
+
+import java.util.Collection;
+import java.util.Optional;
+import java.util.Set;
+import org.junit.jupiter.api.BeforeEach;
+import tech.pegasys.teku.bls.BLSPublicKey;
+import tech.pegasys.teku.bls.BLSSignature;
+import tech.pegasys.teku.core.signatures.Signer;
+import tech.pegasys.teku.datastructures.state.ForkInfo;
+import tech.pegasys.teku.datastructures.util.DataStructureUtil;
+import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.async.StubAsyncRunner;
+import tech.pegasys.teku.metrics.StubMetricsSystem;
+import tech.pegasys.teku.validator.api.ValidatorApiChannel;
+import tech.pegasys.teku.validator.client.duties.AttestationProductionDuty;
+import tech.pegasys.teku.validator.client.duties.ValidatorDutyFactory;
 
 @SuppressWarnings("FutureReturnValueIgnored")
 public abstract class AbstractDutySchedulerTest {
@@ -60,7 +61,10 @@ public abstract class AbstractDutySchedulerTest {
         .thenReturn(completedFuture(Optional.of(emptyList())));
     when(dutyFactory.createAttestationProductionDuty(any()))
         .thenReturn(mock(AttestationProductionDuty.class));
+    final SafeFuture<BLSSignature> rejectAggregationSignature =
+        SafeFuture.failedFuture(new UnsupportedOperationException("This test ignores aggregation"));
+    when(validator1Signer.signAggregationSlot(any(), any())).thenReturn(rejectAggregationSignature);
+    when(validator2Signer.signAggregationSlot(any(), any())).thenReturn(rejectAggregationSignature);
     when(forkProvider.getForkInfo()).thenReturn(completedFuture(fork));
   }
-
 }
