+++ validator/client/src/test/java/tech/pegasys/teku/validator/client/BlockDutySchedulerTest.java
@@ -13,20 +13,6 @@
 
 package tech.pegasys.teku.validator.client;
 
-import org.junit.jupiter.api.Test;
-import tech.pegasys.teku.infrastructure.async.SafeFuture;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.metrics.StubMetricsSystem;
-import tech.pegasys.teku.validator.api.ValidatorDuties;
-import tech.pegasys.teku.validator.api.ValidatorTimingChannel;
-import tech.pegasys.teku.validator.client.duties.AttestationProductionDuty;
-import tech.pegasys.teku.validator.client.duties.BlockProductionDuty;
-import tech.pegasys.teku.validator.client.duties.ScheduledDuties;
-
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-
 import static java.util.Collections.emptyList;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
@@ -39,6 +25,18 @@ import static tech.pegasys.teku.datastructures.util.BeaconStateUtil.compute_star
 import static tech.pegasys.teku.infrastructure.async.SafeFuture.completedFuture;
 import static tech.pegasys.teku.infrastructure.unsigned.UInt64.ONE;
 import static tech.pegasys.teku.infrastructure.unsigned.UInt64.ZERO;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import org.junit.jupiter.api.Test;
+import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.metrics.StubMetricsSystem;
+import tech.pegasys.teku.validator.api.ValidatorDuties;
+import tech.pegasys.teku.validator.api.ValidatorTimingChannel;
+import tech.pegasys.teku.validator.client.duties.BlockProductionDuty;
+import tech.pegasys.teku.validator.client.duties.ScheduledDuties;
 
 public class BlockDutySchedulerTest extends AbstractDutySchedulerTest {
   private final BlockDutyScheduler dutyScheduler =
@@ -108,59 +106,6 @@ public class BlockDutySchedulerTest extends AbstractDutySchedulerTest {
     // Execute
     dutyScheduler.onBlockProductionDue(blockProposerSlot);
     verify(blockCreationDuty).performDuty();
-  }
-
-  @Test
-  public void shouldNotScheduleAttestationDutiesTwice() {
-    final UInt64 attestationSlot = UInt64.valueOf(5);
-    final int validator1Index = 5;
-    final int validator1Committee = 3;
-    final int validator1CommitteePosition = 9;
-    final int validator2Index = 6;
-    final int validator2Committee = 4;
-    final int validator2CommitteePosition = 8;
-    final ValidatorDuties validator1Duties =
-        ValidatorDuties.withDuties(
-            VALIDATOR1_KEY,
-            validator1Index,
-            validator1Committee,
-            validator1CommitteePosition,
-            0,
-            emptyList(),
-            attestationSlot);
-    final ValidatorDuties validator2Duties =
-        ValidatorDuties.withDuties(
-            VALIDATOR2_KEY,
-            validator2Index,
-            validator2Committee,
-            validator2CommitteePosition,
-            0,
-            emptyList(),
-            attestationSlot);
-    when(validatorApiChannel.getDuties(eq(ZERO), any()))
-        .thenReturn(completedFuture(Optional.of(List.of(validator1Duties, validator2Duties))));
-
-    final AttestationProductionDuty attestationDuty = mock(AttestationProductionDuty.class);
-    when(attestationDuty.performDuty()).thenReturn(new SafeFuture<>());
-    when(dutyFactory.createAttestationProductionDuty(attestationSlot)).thenReturn(attestationDuty);
-
-    // Load duties
-    dutyScheduler.onSlot(compute_start_slot_at_epoch(ZERO));
-
-    // Both validators should be scheduled to create an attestation in the same slot
-    verify(attestationDuty)
-        .addValidator(
-            validator1, validator1Committee, validator1CommitteePosition, validator1Index);
-    verify(attestationDuty)
-        .addValidator(
-            validator2, validator2Committee, validator2CommitteePosition, validator2Index);
-
-    // Execute
-    dutyScheduler.onBlockImportedForSlot(attestationSlot);
-    verify(attestationDuty, never()).performDuty();
-
-    dutyScheduler.onAttestationCreationDue(attestationSlot);
-    verifyNoMoreInteractions(attestationDuty);
   }
 
   @Test
